7-llm/pyproject.toml
[project]
name = "7-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "openai>=2.8.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupytext>=1.17.3",
]
